refactor things into class-api
