--- src/libs/context-api/controller/index.ts
@@ -1,1 +1,0 @@
-export * from './context-base.class.js';
