+++ src/libs/context-api/index.ts
@@ -1,5 +1,4 @@
 export * from './consume/index.js';
-export * from './controller/index.js';
 export * from './debug/context-data.function.js';
 export * from './provide/index.js';
 export * from './token/index.js';
